refactoring: splitting the AbstractOperator into smaller chunks
