+++ src/main/java/io/radanalytics/operator/common/CustomResourceWatcher.java
@@ -1,236 +1,86 @@
 package io.radanalytics.operator.common;
 
-import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
-import io.fabric8.kubernetes.api.builder.Function;
-import io.fabric8.kubernetes.api.model.ConfigMap;
-import io.fabric8.kubernetes.api.model.ConfigMapList;
-import io.fabric8.kubernetes.api.model.DoneableConfigMap;
 import io.fabric8.kubernetes.api.model.apiextensions.CustomResourceDefinition;
-import io.fabric8.kubernetes.client.*;
-import io.fabric8.kubernetes.client.dsl.MixedOperation;
-import io.fabric8.kubernetes.client.dsl.Resource;
-import io.fabric8.kubernetes.client.dsl.Watchable;
-import io.fabric8.kubernetes.internal.KubernetesDeserializer;
-import io.radanalytics.operator.Entrypoint;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import io.fabric8.kubernetes.client.KubernetesClient;
+import io.radanalytics.operator.common.crd.InfoClass;
 
-import java.util.Map;
 import java.util.concurrent.CompletableFuture;
 import java.util.function.BiConsumer;
-import java.util.function.Predicate;
+import java.util.function.Function;
 
-import static io.radanalytics.operator.common.AnsiColors.*;
-
-public abstract class AbstractCustomResourceWatcher<T extends EntityInfo> extends AbstractWatcher<T> {
-
-    protected static final Logger log = LoggerFactory.getLogger(AbstractCustomResourceWatcher.class.getName());
-
-    private final boolean isCrd;
-    private final String namespace;
-    private final String entityName;
-    private final KubernetesClient client;
-    private final CustomResourceDefinition crd;
-    private final Map<String, String> selector;
-
-    private final BiConsumer<T, String> onAdd;
-    private final BiConsumer<T, String> onDelete;
-    private final BiConsumer<T, String> onModify;
-
-    private final Predicate<ConfigMap> isSupported;
-    private final Function<ConfigMap, T> convert;
-    private final Function<InfoClass, T> convertCr;
-
-    private volatile Watch watch;
-    protected volatile boolean fullReconciliationRun = false;
+public class CustomResourceWatcher<T extends EntityInfo> extends AbstractWatcher<T> {
 
     // use via builder
-    private AbstractCustomResourceWatcher(boolean isCrd, String namespace, String entityName, KubernetesClient client,
-                                          CustomResourceDefinition crd, Map<String, String> selector, BiConsumer<T, String> onAdd,
-                                          BiConsumer<T, String> onDelete, BiConsumer<T, String> onModify, Predicate<ConfigMap> isSupported,
-                                          Function<ConfigMap, T> convert, Function<InfoClass, T> convertCr) {
-        this.isCrd = isCrd;
-        this.namespace = namespace;
-        this.entityName = entityName;
-        this.client = client;
-        this.crd = crd;
-        this.selector = selector;
-        this.onAdd = onAdd;
-        this.onDelete = onDelete;
-        this.onModify = onModify;
-        this.isSupported = isSupported;
-        this.convert = convert;
-        this.convertCr = convertCr;
+    private CustomResourceWatcher(String namespace,
+                                  String entityName,
+                                  KubernetesClient client,
+                                  CustomResourceDefinition crd,
+                                  BiConsumer<T, String> onAdd,
+                                  BiConsumer<T, String> onDelete,
+                                  BiConsumer<T, String> onModify,
+                                  Function<InfoClass, T> convert) {
+        super(true, namespace, entityName, client, crd, null, onAdd, onDelete, onModify, null, null, convert);
     }
 
-    public CompletableFuture<AbstractCustomResourceWatcher> build() {
-        if (isCrd) {
-            createCRDWatch();
-        } else {
-            createConfigMapWatch();
-        }
-        return null;
-    }
+    public static class Builder<T> {
+        private String namespace = "*";
+        private String entityName;
+        private KubernetesClient client;
+        private CustomResourceDefinition crd;
 
-    private CompletableFuture<Watch> createConfigMapWatch() {
-        CompletableFuture<Watch> cf = CompletableFuture.supplyAsync(() -> {
-            MixedOperation<ConfigMap, ConfigMapList, DoneableConfigMap, Resource<ConfigMap, DoneableConfigMap>> aux = client.configMaps();
+        private BiConsumer<T, String> onAdd;
+        private BiConsumer<T, String> onDelete;
+        private BiConsumer<T, String> onModify;
+        private Function<InfoClass, T> convert;
 
-            final boolean inAllNs = "*".equals(namespace);
-            Watchable<Watch, Watcher<ConfigMap>> watchable = inAllNs ? aux.inAnyNamespace().withLabels(selector) : aux.inNamespace(namespace).withLabels(selector);
-            Watch watch = watchable.watch(new Watcher<ConfigMap>() {
-                @Override
-                public void eventReceived(Action action, ConfigMap cm) {
-                    if (isSupported.test(cm)) {
-                        log.info("ConfigMap \n{}\n in namespace {} was {}", cm, namespace, action);
-                        T entity = convert.apply(cm);
-                        if (entity == null) {
-                            log.error("something went wrong, unable to parse {} definition", entityName);
-                        }
-                        if (action.equals(Action.ERROR)) {
-                            log.error("Failed ConfigMap {} in namespace{} ", cm, namespace);
-                        } else {
-                            handleAction(action, entity, inAllNs ? cm.getMetadata().getNamespace() : namespace);
-                        }
-                    } else {
-                        log.error("Unknown CM kind: {}", cm.toString());
-                    }
-                }
-
-                @Override
-                public void onClose(KubernetesClientException e) {
-                    if (e != null) {
-                        log.error("Watcher closed with exception in namespace {}", namespace, e);
-                        recreateWatcher();
-                    } else {
-                        log.info("Watcher closed in namespace {}", namespace);
-                    }
-                }
-            });
-            return watch;
-        }, Entrypoint.EXECUTORS);
-        cf.thenApply(w -> {
-            log.info("ConfigMap watcher running for labels {}", selector);
-            return w;
-        }).exceptionally(e -> {
-            log.error("ConfigMap watcher failed to start", e.getCause());
-            return null;
-        });
-        return cf;
-    }
-
-    public static class InfoClass<U> extends CustomResource {
-        private U spec;
-
-        public U getSpec() {
-            return spec;
+        public Builder<T> withNamespace(String namespace) {
+            this.namespace = namespace;
+            return this;
         }
 
-        public void setSpec(U spec) {
-            this.spec = spec;
+        public Builder<T> withEntityName(String entityName) {
+            this.entityName = entityName;
+            return this;
+        }
+
+        public Builder<T> withClient(KubernetesClient client) {
+            this.client = client;
+            return this;
+        }
+
+        public Builder<T> withCrd(CustomResourceDefinition crd) {
+            this.crd = crd;
+            return this;
+        }
+
+        public Builder<T> withOnAdd(BiConsumer<T, String> onAdd) {
+            this.onAdd = onAdd;
+            return this;
+        }
+
+        public Builder<T> withOnDelete(BiConsumer<T, String> onDelete) {
+            this.onDelete = onDelete;
+            return this;
+        }
+
+        public Builder<T> withOnModify(BiConsumer<T, String> onModify) {
+            this.onModify = onModify;
+            return this;
+        }
+
+        public Builder<T> withConvert(Function<InfoClass, T> convert) {
+            this.convert = convert;
+            return this;
+        }
+
+        public CustomResourceWatcher build() {
+            return new CustomResourceWatcher(namespace, entityName, client, crd, onAdd, onDelete, onModify, convert);
         }
     }
 
-    public static class InfoClassDoneable<S> extends CustomResourceDoneable<InfoClass<S>> {
-        public InfoClassDoneable(InfoClass<S> resource, Function function) {
-            super(resource, function);
-        }
-    }
-
-    @JsonDeserialize(using = KubernetesDeserializer.class)
-    public class InfoList<V> extends CustomResourceList<InfoClass<V>> {
-    }
-
-    private CompletableFuture<Watch> createCRDWatch() {
-        CompletableFuture<Watch> cf = CompletableFuture.supplyAsync(() -> {
-            MixedOperation<InfoClass, InfoList, InfoClassDoneable, Resource<InfoClass, InfoClassDoneable>> aux =
-                    client.customResources(crd, InfoClass.class, InfoList.class, InfoClassDoneable.class);
-
-            final boolean inAllNs = "*".equals(namespace);
-            Watchable<Watch, Watcher<InfoClass>> watchable = inAllNs ? aux.inAnyNamespace() : aux.inNamespace(namespace);
-            Watch watch = watchable.watch(new Watcher<InfoClass>() {
-                @Override
-                public void eventReceived(Action action, InfoClass info) {
-                    log.info("Custom resource \n{}\n in namespace {} was {}", info, namespace, action);
-                    T entity = convertCr.apply(info);
-                    if (entity == null) {
-                        log.error("something went wrong, unable to parse {} definition", entityName);
-                    }
-                    if (action.equals(Action.ERROR)) {
-                        log.error("Failed Custom resource {} in namespace{} ", info, namespace);
-                    } else {
-                        handleAction(action, entity, inAllNs ? info.getMetadata().getNamespace() : namespace);
-                    }
-                }
-
-                @Override
-                public void onClose(KubernetesClientException e) {
-                    if (e != null) {
-                        log.error("Watcher closed with exception in namespace {}", namespace, e);
-                        recreateWatcher();
-                    } else {
-                        log.info("Watcher closed in namespace {}", namespace);
-                    }
-                }
-            });
-            return watch;
-        }, Entrypoint.EXECUTORS);
-        cf.thenApply(w -> {
-            log.info("CustomResource watcher running for kinds {}", entityName);
-            return w;
-        }).exceptionally(e -> {
-            log.error("CustomResource watcher failed to start", e.getCause());
-            return null;
-        });
-        return cf;
-    }
-
-    private void recreateWatcher() {
-        CompletableFuture<Watch> configMapWatch = isCrd ? createCRDWatch(this.crd): createConfigMapWatch();
-        final String crdOrCm = isCrd ? "CustomResource" : "ConfigMap";
-        configMapWatch.thenApply(res -> {
-            log.info("{} watch recreated in namespace {}", crdOrCm, namespace);
-            this.watch = res;
-            return res;
-        }).exceptionally(e -> {
-            log.error("Failed to recreate {} watch in namespace {}", crdOrCm, namespace);
-            return null;
-        });
-    }
-
-    private void handleAction(Watcher.Action action, T entity, String ns) {
-        if (!fullReconciliationRun) {
-            return;
-        }
-        String name = entity.getName();
-        try {
-            switch (action) {
-                case ADDED:
-                    log.info("{}creating{} {}:  \n{}\n", gr(), xx(), entityName, name);
-                    onAdd.accept(entity, ns);
-                    log.info("{} {} has been  {}created{}", entityName, name, gr(), xx());
-                    break;
-                case DELETED:
-                    log.info("{}deleting{} {}:  \n{}\n", gr(), xx(), entityName, name);
-                    onDelete.accept(entity, ns);
-                    log.info("{} {} has been  {}deleted{}", entityName, name, gr(), xx());
-                    break;
-                case MODIFIED:
-                    log.info("{}modifying{} {}:  \n{}\n", gr(), xx(), entityName, name);
-                    onModify.accept(entity, ns);
-                    log.info("{} {} has been  {}modified{}", entityName, name, gr(), xx());
-                    break;
-                default:
-                    log.error("Unknown action: {} in namespace {}", action, namespace);
-            }
-        } catch (Exception e) {
-            log.warn("{}Error{} when reacting on event, cause: {}", re(), xx(), e.getMessage());
-            e.printStackTrace();
-        }
-    }
-
-    public void setFullReconciliationRun(boolean fullReconciliationRun) {
-        this.fullReconciliationRun = fullReconciliationRun;
+    @Override
+    public CompletableFuture<CustomResourceWatcher<T>> watch() {
+        return createCustomResourceWatch().thenApply(watch -> this);
     }
 }
