+++ src/main/java/io/radanalytics/operator/common/crd/InfoClassDoneable.java
@@ -1,11 +1,11 @@
-package io.radanalytics.operator.common;
+package io.radanalytics.operator.common.crd;
 
 import io.fabric8.kubernetes.api.builder.Function;
 import io.fabric8.kubernetes.client.CustomResourceDoneable;
 
 
-public class InfoClassDoneable<S> extends CustomResourceDoneable<AbstractWatcher.InfoClass<S>> {
-    public InfoClassDoneable(AbstractWatcher.InfoClass<S> resource, Function function) {
+public class InfoClassDoneable<S> extends CustomResourceDoneable<InfoClass<S>> {
+    public InfoClassDoneable(InfoClass<S> resource, Function function) {
         super(resource, function);
     }
 }
